+++ packages/@expo/fingerprint/src/ProjectWorkflow.ts
@@ -8,6 +8,7 @@ import path from 'path';
 
 import { resolveExpoConfigPluginsPackagePath } from './ExpoResolver';
 import type { Platform, ProjectWorkflow } from './Fingerprint.types';
+import { collator } from './utils/Collator';
 import { isIgnoredPathWithMatchObjects, pathExistsAsync, toPosixPath } from './utils/Path';
 
 /**
@@ -163,7 +164,7 @@ class Ignore {
       })
     )
       // ensure that parent dir is before child directories
-      .sort((a, b) => a.length - b.length && a.localeCompare(b));
+      .sort((a, b) => a.length - b.length && collator.compare(a, b));
 
     const ignoreMapping = await Promise.all(
       ignoreFilePaths.map(async (filePath) => {