+++ packages/@expo/fingerprint/src/Sort.ts
@@ -1,4 +1,5 @@
 import type { HashSource } from './Fingerprint.types';
+import { collator } from './utils/Collator';
 
 export function sortSources<T extends HashSource>(sources: T[]): T[] {
   return sources.sort(compareSource);
@@ -25,17 +26,17 @@ export function compareSource(a: HashSource, b: HashSource): number {
     if (a.type === 'file' && b.type === 'file') {
       const aValue = a.overrideHashKey ?? a.filePath;
       const bValue = b.overrideHashKey ?? b.filePath;
-      return aValue.localeCompare(bValue);
+      return collator.compare(aValue, bValue);
     } else if (a.type === 'dir' && b.type === 'dir') {
       const aValue = a.overrideHashKey ?? a.filePath;
       const bValue = b.overrideHashKey ?? b.filePath;
-      return aValue.localeCompare(bValue);
+      return collator.compare(aValue, bValue);
     } else if (a.type === 'contents' && b.type === 'contents') {
-      return a.id.localeCompare(b.id);
+      return collator.compare(a.id, b.id);
     } else if (a.type === 'package' && b.type === 'package') {
       const aValue = a.overrideHashKey ?? `${a.name}@${a.version}`;
       const bValue = b.overrideHashKey ?? `${b.name}@${b.version}`;
-      return aValue.localeCompare(bValue);
+      return collator.compare(aValue, bValue);
     }
   }
   return typeResult;